Extract some code into file RemoveBackEdges
